[#408] Extract shared compiler support module
CharArrayCharStream (6 identical copies) and ParsingUtils (4 copies)
move to new :plugins:compiler:compiler-shared module. Application
depends on it so the jar lands in dist lib/, where plugin manifests
resolve Class-Path. ADR-0008 records the boundary decision.

Not unified: per-plugin exceptions and ParserErrorListener (divergent
types/formats), FlagsModel (different semantics), Breakpoint (trivial),
terminal DisplayCanvas (too divergent).
