+++ plugins/compiler/as-8080/src/main/java/net/emustudio/plugins/compiler/as8080/ast/expr/ExprString.java
@@ -11,7 +11,7 @@ import org.antlr.v4.runtime.Token;
 
 import java.util.Objects;
 
-import static net.emustudio.plugins.compiler.as8080.ParsingUtils.parseLitString;
+import static net.emustudio.plugins.compiler.shared.ParsingUtils.parseLitString;
 
 public class ExprString extends Node {
     public final String string;
