+++ plugins/compiler/as-8080/src/main/java/net/emustudio/plugins/compiler/as8080/ast/pseudo/PseudoInclude.java
@@ -3,7 +3,7 @@
 package net.emustudio.plugins.compiler.as8080.ast.pseudo;
 
 import net.emustudio.emulib.plugins.compiler.SourceCodePosition;
-import net.emustudio.plugins.compiler.as8080.ParsingUtils;
+import net.emustudio.plugins.compiler.shared.ParsingUtils;
 import net.emustudio.plugins.compiler.as8080.ast.Node;
 import net.emustudio.plugins.compiler.as8080.visitors.NodeVisitor;
 import org.antlr.v4.runtime.Token;
