+++ plugins/compiler/as-8080/src/main/java/net/emustudio/plugins/compiler/as8080/visitors/CheckDeclarationsVisitor.java
@@ -9,7 +9,7 @@ import net.emustudio.plugins.compiler.as8080.ast.pseudo.*;
 import java.util.*;
 
 import static net.emustudio.plugins.compiler.as8080.CompileError.*;
-import static net.emustudio.plugins.compiler.as8080.ParsingUtils.normalizeId;
+import static net.emustudio.plugins.compiler.shared.ParsingUtils.normalizeId;
 
 /**
  * Checks if all declarations are valid:
