+++ plugins/compiler/as-8080/src/main/java/net/emustudio/plugins/compiler/as8080/visitors/CreateExprVisitor.java
@@ -5,7 +5,7 @@ package net.emustudio.plugins.compiler.as8080.visitors;
 import net.emustudio.emulib.plugins.compiler.SourceCodePosition;
 import net.emustudio.plugins.compiler.as8080.As8080Parser;
 import net.emustudio.plugins.compiler.as8080.As8080ParserBaseVisitor;
-import net.emustudio.plugins.compiler.as8080.ParsingUtils;
+import net.emustudio.plugins.compiler.shared.ParsingUtils;
 import net.emustudio.plugins.compiler.as8080.ast.Node;
 import net.emustudio.plugins.compiler.as8080.ast.expr.*;
 import org.antlr.v4.runtime.Token;
