+++ plugins/compiler/as-8080/src/main/java/net/emustudio/plugins/compiler/as8080/visitors/EvaluateExprVisitor.java
@@ -17,7 +17,7 @@ import java.util.stream.Collectors;
 
 import static net.emustudio.plugins.compiler.as8080.CompileError.ambiguousExpression;
 import static net.emustudio.plugins.compiler.as8080.CompileError.expressionIsBiggerThanExpected;
-import static net.emustudio.plugins.compiler.as8080.ParsingUtils.normalizeId;
+import static net.emustudio.plugins.compiler.shared.ParsingUtils.normalizeId;
 
 /**
  * The goal is to replace all Expr* with Evaluated
