+++ plugins/compiler/as-8080/src/main/java/net/emustudio/plugins/compiler/as8080/visitors/SortMacroArgumentsVisitor.java
@@ -15,7 +15,7 @@ import java.util.List;
 import java.util.Map;
 
 import static net.emustudio.plugins.compiler.as8080.CompileError.macroArgumentsDoNotMatch;
-import static net.emustudio.plugins.compiler.as8080.ParsingUtils.normalizeId;
+import static net.emustudio.plugins.compiler.shared.ParsingUtils.normalizeId;
 
 /**
  * Connects macro parameters with arguments.
