+++ plugins/compiler/as-ssem/build.gradle
@@ -19,6 +19,7 @@ dependencies {
   implementation libs.antlrRuntime
 
   implementation libs.emuLib
+  implementation project(":plugins:compiler:compiler-shared")
   implementation libs.slf4JApi
   implementation libs.jcipAnnotations
