+++ plugins/compiler/as-z80/build.gradle
@@ -19,6 +19,7 @@ dependencies {
   implementation libs.antlrRuntime
 
   implementation libs.emuLib
+  implementation project(":plugins:compiler:compiler-shared")
   implementation libs.slf4JApi
 
   testImplementation libs.cpuTestSuite
