+++ plugins/compiler/as-z80/src/main/java/net/emustudio/plugins/compiler/asZ80/visitors/CheckDeclarationsVisitor.java
@@ -9,7 +9,7 @@ import net.emustudio.plugins.compiler.asZ80.ast.pseudo.*;
 import java.util.*;
 
 import static net.emustudio.plugins.compiler.asZ80.CompileError.*;
-import static net.emustudio.plugins.compiler.asZ80.ParsingUtils.normalizeId;
+import static net.emustudio.plugins.compiler.shared.ParsingUtils.normalizeId;
 
 /**
  * Checks if all declarations are valid:
