+++ plugins/compiler/as-z80/src/main/java/net/emustudio/plugins/compiler/asZ80/visitors/SortMacroArgumentsVisitor.java
@@ -15,7 +15,7 @@ import java.util.List;
 import java.util.Map;
 
 import static net.emustudio.plugins.compiler.asZ80.CompileError.macroArgumentsDoNotMatch;
-import static net.emustudio.plugins.compiler.asZ80.ParsingUtils.normalizeId;
+import static net.emustudio.plugins.compiler.shared.ParsingUtils.normalizeId;
 
 /**
  * Connects macro parameters with arguments.
