+++ plugins/compiler/compiler-shared/build.gradle
@@ -1,0 +1,25 @@
+/* SPDX-FileCopyrightText: 2006-2026 Peter Jakubčo
+   SPDX-License-Identifier: GPL-3.0-or-later */
+
+plugins {
+  id 'java'
+  id 'idea'
+  id 'com.adarshr.test-logger' version '4.0.0'
+  id 'jacoco'
+}
+
+dependencies {
+  implementation libs.antlrRuntime
+
+  testImplementation libs.junit
+}
+
+apply from: "$rootDir/gradle/conventions/java-conventions.gradle"
+
+test {
+  finalizedBy jacocoTestReport
+}
+
+jacocoTestReport {
+  dependsOn test
+}
