+++ plugins/compiler/ramc-ram/src/main/java/net/emustudio/plugins/compiler/ram/ProgramParser.java
@@ -13,7 +13,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 
-import static net.emustudio.plugins.compiler.ram.ParsingUtils.*;
+import static net.emustudio.plugins.compiler.shared.ParsingUtils.*;
 import static net.emustudio.plugins.compiler.ram.RAMParser.*;
 
 public class ProgramParser extends RAMParserBaseVisitor<Program> {
